refactor: ClientNotFoundException 도메인 exception 패키지로 이동
