+++ src/main/java/com/map/gaja/client/apllication/ClientAccessVerifyService.java
@@ -3,7 +3,7 @@ package com.map.gaja.client.apllication;
 import com.map.gaja.group.application.GroupAccessVerifyService;
 import com.map.gaja.client.infrastructure.repository.ClientQueryRepository;
 import com.map.gaja.client.presentation.dto.ClientAccessCheckDto;
-import com.map.gaja.client.presentation.exception.ClientNotFoundException;
+import com.map.gaja.client.domain.exception.ClientNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
