+++ src/test/java/com/map/gaja/client/apllication/ClientAccessCheckServiceTest.java
@@ -6,7 +6,7 @@ import com.map.gaja.client.domain.model.Client;
 import com.map.gaja.client.domain.model.ClientAddress;
 import com.map.gaja.client.domain.model.ClientLocation;
 import com.map.gaja.client.presentation.dto.ClientAccessCheckDto;
-import com.map.gaja.client.presentation.exception.ClientNotFoundException;
+import com.map.gaja.client.domain.exception.ClientNotFoundException;
 import com.map.gaja.user.domain.model.Authority;
 import com.map.gaja.user.domain.model.User;
 import org.junit.jupiter.api.BeforeEach;
